+++ innopoints/schemas/__init__.py
@@ -2,5 +2,6 @@ from .project import *
 from .account import *
 from .activity import *
 from .application import *
+from .product import *
 
 from .schemas import *  # temporary until moved all schemas
